+++ drizzle/0004_fixed_electro.sql
@@ -1,0 +1,4 @@
+ALTER TYPE "public"."payment_provider" ADD VALUE 'midtrans';--> statement-breakpoint
+ALTER TABLE "payments" ADD COLUMN "gross_amount" integer;--> statement-breakpoint
+ALTER TABLE "payments" ADD COLUMN "qr_string" text;--> statement-breakpoint
+ALTER TABLE "payments" ADD COLUMN "expires_at" timestamp with time zone;